Factor out the EnterName component
